Formalized utility for moving Firebase DB events to Exector threads.
ExecutorValueEventListener lets you easily create a Firebase database listener that proxies the handling of callbacks from the main thread to a thread determined by an Executor.  This helps you offload expensive parsing and computation so that your app can render smoothly at the same time.  The strategy here is to broadcast the parsed POJOs through an event bus as they arrive.