Save connections and data when app isn't visible.
This trick will cut down on the number of concurrent connections to Firebase from an Android app when it's not being used.  It may also reduce the amount of data being used by the app, if it retains persistent listeners or uses keepSynced().  See the comment on GoOfflineWhenInvisible for more information.